+++ market/build.rs
@@ -1,10 +1,9 @@
-// use serde::{Deserialize, Serialize};
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let proto = "market/market.proto";
+  let proto = "proto/market.proto";
 
-    tonic_build::configure()
-        .type_attribute("User", "#[derive(serde::Deserialize, serde::Serialize)]")
-        .compile(&[proto], &["."])?;
+  tonic_build::configure()
+      .type_attribute("User", "#[derive(serde::Deserialize, serde::Serialize)]")
+      .compile(&[proto], &["../"])?;
 
-    Ok(())
+  Ok(())
 }
