+++ peernode/build.rs
@@ -1,4 +1,9 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::compile_protos("../proto/market.proto")?;
-    Ok(())
+  let proto = "proto/market.proto";
+
+  tonic_build::configure()
+      .type_attribute("User", "#[derive(serde::Deserialize, serde::Serialize)]")
+      .compile(&[proto], &["../"])?;
+
+  Ok(())
 }
